+++ src/EditorFeatures/CSharpTest/SimplifyTypeNames/SimplifyTypeNamesTests_FixAllTests.cs
@@ -590,93 +590,6 @@ class ProgramB3
             await TestInRegularAndScriptAsync(input, expected);
         }
 
-        [Fact]
-        [Trait(Traits.Feature, Traits.Features.CodeActionsQualifyMemberAccess)]
-        [Trait(Traits.Feature, Traits.Features.CodeActionsFixAllOccurrences)]
-        public async Task TestFixAllInSolution_RemoveMemberAccessQualification()
-        {
-            var input = @"
-<Workspace>
-    <Project Language=""C#"" AssemblyName=""Assembly1"" CommonReferences=""true"">
-        <Document>
-using System;
-
-class C
-{
-    int Property { get; set; }
-    int OtherProperty { get; set; }
-
-    void M()
-    {
-        {|FixAllInSolution:this.Property|} = 1;
-        var x = this.OtherProperty;
-    }
-}
-        </Document>
-        <Document>
-using System;
-
-class D
-{
-    string StringProperty { get; set; }
-    int field;
-
-    void N()
-    {
-        this.StringProperty = string.Empty;
-        this.field = 0; // ensure qualification isn't removed
-    }
-}
-        </Document>
-    </Project>
-</Workspace>";
-
-            var expected = @"
-<Workspace>
-    <Project Language=""C#"" AssemblyName=""Assembly1"" CommonReferences=""true"">
-        <Document>
-using System;
-
-class C
-{
-    int Property { get; set; }
-    int OtherProperty { get; set; }
-
-    void M()
-    {
-        Property = 1;
-        var x = OtherProperty;
-    }
-}
-        </Document>
-        <Document>
-using System;
-
-class D
-{
-    string StringProperty { get; set; }
-    int field;
-
-    void N()
-    {
-        StringProperty = string.Empty;
-        this.field = 0; // ensure qualification isn't removed
-    }
-}
-        </Document>
-    </Project>
-</Workspace>";
-
-            var options = OptionsSet(
-                SingleOption(CodeStyleOptions.QualifyPropertyAccess, false, NotificationOption.Suggestion),
-                SingleOption(CodeStyleOptions.QualifyFieldAccess, true, NotificationOption.Suggestion));
-
-            await TestInRegularAndScriptAsync(
-                initialMarkup: input,
-                expectedMarkup: expected,
-                options: options);
-        }
-
         #endregion
     }
 }
