+++ src/EditorFeatures/VisualBasicTest/SimplifyTypeNames/SimplifyTypeNamesTests_FixAllTests.vb
@@ -502,82 +502,5 @@ End Class]]>
 
             Await TestInRegularAndScriptAsync(input, expected)
         End Function
-
-        <Fact>
-        <Trait(Traits.Feature, Traits.Features.CodeActionsQualifyMemberAccess)>
-        <Trait(Traits.Feature, Traits.Features.CodeActionsFixAllOccurrences)>
-        Public Async Function TestFixAllInSolution_RemoveMemberAccessQualification() As Task
-            Dim input =
-<Workspace>
-    <Project Language="Visual Basic" AssemblyName="Assembly1" CommonReferences="true">
-        <Document><![CDATA[
-Imports System
-
-Class C
-    Property SomeProperty As Integer
-    Property OtherProperty As Integer
-
-    Sub M()
-        {|FixAllInSolution:Me.SomeProperty|} = 1
-        Dim x = Me.OtherProperty
-    End Sub
-End Class]]>
-        </Document>
-        <Document><![CDATA[
-Imports System
-
-Class D
-    Property StringProperty As String
-    field As Integer
-
-    Sub N()
-        Me.StringProperty = String.Empty
-        Me.field = 0 ' ensure qualification isn't removed
-    End Sub
-End Class]]>
-        </Document>
-    </Project>
-</Workspace>.ToString()
-
-            Dim expected =
-<Workspace>
-    <Project Language="Visual Basic" AssemblyName="Assembly1" CommonReferences="true">
-        <Document><![CDATA[
-Imports System
-
-Class C
-    Property SomeProperty As Integer
-    Property OtherProperty As Integer
-
-    Sub M()
-        SomeProperty = 1
-        Dim x = OtherProperty
-    End Sub
-End Class]]>
-        </Document>
-        <Document><![CDATA[
-Imports System
-
-Class D
-    Property StringProperty As String
-    field As Integer
-
-    Sub N()
-        StringProperty = String.Empty
-        Me.field = 0 ' ensure qualification isn't removed
-    End Sub
-End Class]]>
-        </Document>
-    </Project>
-</Workspace>.ToString()
-
-            Dim options = OptionsSet(
-                SingleOption(CodeStyleOptions.QualifyPropertyAccess, False, NotificationOption.Suggestion),
-                SingleOption(CodeStyleOptions.QualifyFieldAccess, True, NotificationOption.Suggestion))
-            Await TestInRegularAndScriptAsync(
-                initialMarkup:=input,
-                expectedMarkup:=expected,
-                options:=options)
-        End Function
     End Class
 End Namespace
